Fix qrc case when resolving url to local path
A url containing a qrc path ("qrc:/content/fonts") is not a local
file and we have to handle this as a special case.

If the url is empty we skip any font loading, since we do not want
to scan complete disks.

I also replaced qCInfo with qCDebug.

Task-number: QDS-6912
Change-Id: Iae8292547abc1811243ea33a63550e54aa02a6a8
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>