Add instructions to generate a PySide package
This includes a new directory with the process to generate
a Python Package (wheel) compatible with PySide6, and including
the special Quick components that QtDS project might contain.

The current version assumes the main project will be built and
installed on a 'install' directory at the root of the repo,
so files can be copied around for the wheel structure.

The instructions can be found on the README, but in summary
one needs to run a 'configure.py' script first, then generate
the package with a 'python -m build -w'

Fixes: PYSIDE-2373
Change-Id: Ic6f8da02f991107180b9a5c60cce944275ca2841
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>