ColorizedImage: Change the way to colorize
ColorOverlay would be used as a sibling of the image to fix an issue
which the effect was not updated in the 2D view.
It led to changing base QML type from Image to Item, and some properties
in the Properties panel were not supported in ColorizedImage type, so
the unsupported fields in the panel would be removed.

Task-number: QDS-11902
Change-Id: I80770358ba6ae44d79c29917cbba68fe462fcfa4
Reviewed-by: Yasser Grimes <yasser.grimes@qt.io>
Reviewed-by: Gatis Paeglis <gatis.paeglis@qt.io>
Reviewed-by: Aleksei German <aleksei.german@qt.io>